% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ices.cod.R
\name{addIces}
\alias{addIces}
\title{Adds ICES assessment results to existing plots}
\usage{
addIces(stock, col = "darkgrey", lwd = 2, lty = c(2, 1, 2),
  what = "FFmsy", mult = 1)
}
\arguments{
\item{stock}{The stock code}

\item{col}{color of the line}

\item{lwd}{line thickness}

\item{lty}{line type}

\item{what}{chr, which column of the result table to plot, see Details.}

\item{mult}{numeric, multiplier of the data (e.g. 1000 for converting tonnes to kg)}
}
\value{
This function is used for its sidefect. It returns \code{NULL} invisibly
}
\description{
Plots lines of asessment results, e.g. F, SSB, etc. to existing plots
}
\note{
If no active plot exists produces an error
}

