% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addVersion}
\alias{addVersion}
\title{Adds version watermark to a plot}
\usage{
addVersion(cex = 0.5, col = "#12345655")
}
\arguments{
\item{cex}{numeric, controls the font size}

\item{col}{A color code or name}
}
\value{
Invisible NULL
}
\description{
Adds the version and, if installed from Github, the first 10 characters from
SHA commit code
}
\examples{
hist(simulateData3()$sample)
addVersion()
}
\author{
alko
}
\seealso{
"Color specification" in the documentation of \code{\link{par}}
}
\keyword{aplot}

