% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{changeBinsize}
\alias{changeBinsize}
\title{Change the bin size of a weight frequency data.frame}
\usage{
changeBinsize(df, binsize = 10, keepZeros = TRUE, weight.col = "Weight",
  freq.col = "Freq")
}
\arguments{
\item{df}{s6input object}

\item{binsize}{numeric, the bin size in grams}

\item{keepZeros}{logical, if FALSE the bins with zero observations are removed}

\item{weight.col}{The weight column name}

\item{freq.col}{The frequency column name}
}
\value{
data.frame with weight frequencies binned with bin size \code{binsize}
}
\description{
Change the bin size of a weight frequency data.frame
}
\examples{
## Simulate a data set with bin size equal to 100 gr
dat <- simulate(s6params(), binsize = 100)

## Change the bin size to 200 gr
dat <- changeBinsize(dat, binsize = 200)
}

