% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s6params-class.R
\docType{methods}
\name{difference}
\alias{difference}
\alias{difference,s6params,s6params-method}
\title{Difference between two \code{s6params} objects}
\usage{
difference(base, comp)

\S4method{difference}{s6params,s6params}(base, comp)
}
\arguments{
\item{base}{\code{s6params} object. First object}

\item{comp}{\code{s6params} object. Second object}
}
\value{
TRUE if they are the same. If there are differences, a data.frame is returned
with the untransformed parameter values of the two objects, the relative difference (base - comp)
and the percent difference.
}
\description{
Difference between two \code{s6params} objects
}
\author{
alko
}

