% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s6params_class.R
\name{difference}
\alias{difference}
\alias{difference.s6params}
\title{Difference between objects}
\usage{
difference(x, ...)

\method{difference}{s6params}(x, y)
}
\arguments{
\item{x}{First object to compare}
}
\description{
Difference between objects
}
