% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{estimate}
\alias{estimate}
\title{Title}
\usage{
estimate(df, n = 0.75, epsilon_a = 0.8, epsilon_r = 0.1, A = 4.47,
  eta_m = 0.25, a = 0.22, Winf = NULL, sigma = NULL, u = 10,
  sdloga = 0.7, winf.ubound = 2, Wfs = NULL, verbose = FALSE,
  map = list(loga = factor(NA)), random = c(), isSurvey = FALSE,
  eta_S = NULL, usePois = TRUE, totalYield = NULL, ...)
}
\arguments{
\item{df}{}

\item{n}{}

\item{epsilon_a}{}

\item{epsilon_r}{}

\item{A}{}

\item{eta_m}{}

\item{a}{}

\item{Winf}{}

\item{sigma}{}

\item{u}{}

\item{sdloga}{}

\item{winf.ubound}{}

\item{Wfs}{}

\item{verbose}{}

\item{map}{}

\item{random}{}

\item{isSurvey}{}

\item{eta_S}{}

\item{usePois}{}

\item{totalYield}{}

\item{...}{}
}
\description{
Title
}

