% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment.R
\name{fitWL}
\alias{fitWL}
\title{Fit weight-length relationship}
\usage{
fitWL(df, colname.weight = "Weight", colname.length = "Length",
  plotFit = FALSE)
}
\arguments{
\item{df}{\code{data.frame} or \code{DATRASraw} object containing length and weight data}

\item{colname.weight}{chr, the name of the column containing weight data. Only used if df is \code{data.frame}}

\item{colname.length}{chr, the name of the column containing length data. Only used if df is \code{data.frame}}

\item{plotFit}{logical, if TRUE plot the fitted line along with the input data}
}
\value{
A list with a and b parameters and data points used (n)
}
\description{
Fit weight-length relationship
}
\note{
\code{fitWL} return an error if data from more than one species are contained in the DATRASraw object.
}
\examples{
## Simulated data
a <- 0.00058
b <- 3.123
df <- data.frame(lengths = 50:100,  weights =  a * c(50:100) ^ b + rnorm(51))

## Fit weight-length relationship and compare with simulated values of a and b
s6model:::fitWL(df, colname.weight = "we", colname.length = "le")
## a: 0.00057 (sim: 0.00058)
## b: 3.128   (sim: 3.123)

}

