% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{getCor}
\alias{getCor}
\alias{getCor,Parameters-method}
\title{Returns the correlation matrix of the parameters}
\usage{
getCor(object)

\S4method{getCor}{Parameters}(object)
}
\arguments{
\item{object}{A \code{Parameters} object returned by \code{estimateParam}, i.e. having
an attribute hessian}
}
\value{
The correlation matrix
}
\description{
Returns the correlation matrix of the parameters
}
\author{
alko
}

