% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getCor}
\alias{getCor}
\title{Returns the correlation matrix of the parameters}
\usage{
getCor(object)
}
\arguments{
\item{object}{A \code{Parameters} object returned by \code{estimateParam}, i.e. having
an attribute hessian}
}
\value{
The correlation matrix
}
\description{
Returns the correlation matrix of the parameters
}
\author{
alko
}

