% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getDfYears}
\alias{getDfYears}
\title{Get yearly weight frequencies}
\usage{
getDfYears(dat, years = as.numeric(levels(dat[[2]]$Year)), ...)
}
\arguments{
\item{dat}{DATRASraw object}

\item{years}{numeric vector, containing the years to return}

\item{...}{additional arguments to \code{\link{datrasraw2weightfreq}}}
}
\value{
list of \code{data.frame}s, one for each year
}
\description{
For selected years get yearly weight frequencies from DATRAS object
}
\author{
alko
}

