% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{getRandomParameters}
\alias{getRandomParameters}
\alias{getRandomParameters.fixedWinf}
\title{Random model parameters}
\usage{
getRandomParameters(parameter.names = c("A", "n", "eta_m", "eta_F", "a", "Fm",
  "Winf", "epsilon_a", "epsilon_r"), parameter.value = c(4.5, 0.75, 0.25,
  0.05, 0.35, 0.25, 10000, 0.8, 0.1), parameter.sd = c(0, 0, 0, 0.5, 0, 0, 0,
  0, 0), parameter.cv = c(0.5, 0, 0.3, 0, 0.5, 0, 0, 0.1, 0.5),
  parameter.lbound = c(0, 0, 0.01, 0, 0.01, 0, 0, 0, 0),
  parameter.ubound = c(Inf, 0, Inf, Inf, Inf, 0, 0, 0.99, Inf),
  parameter.unif = c(F, F, F, F, F, F, F, F, F), Rrel.gt = -Inf,
  Fmsy.gt = 0)

getRandomParameters.fixedWinf(winf, Rrel.gt = -Inf, Fmsy.gt = 0)
}
\arguments{
\item{parameter.names}{Character vector. Names of the parameters}

\item{parameter.value}{Numeric vector. Parameter mean values}

\item{parameter.sd}{Numeric vector. Standard deviation of the parameters}

\item{parameter.cv}{Numeric vector. Coefficient of variation of log transformed parameters}

\item{parameter.lbound}{Numeric vector. Lower bound of the distributions}

\item{parameter.ubound}{Numeric vector. Upper bound of the distributions}

\item{parameter.unif}{Logical vector. Use uniformly distributed parameters}

\item{Rrel.gt}{Numeric. Relative recruitment constraint. It allows parameters that lead
to Rrel at least equal to Rrel.gt}

\item{Fmsy.gt}{Numeric. Fmsy constraint. It allows parameters that lead
to Fmsy at least equal to Fmsy.gt}

\item{winf}{Numeric. Asymptotic weight}
}
\value{
\code{s6params}

\code{s6params}
}
\description{
Random model parameters with constraints in Fmsy and relative recruitment (Rrel = R/Rmax)

Shorthand function for a specific asymptotic weight
}
\author{
alko

alko
}

