% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{Parameters}
\alias{Parameters}
\alias{Parameters-class}
\title{The Parameters class and constructor}
\description{
A class that contains all model parameters
}
\section{Slots}{

\describe{
  \item{\code{logWinf}:}{Numeric scalar. Asymptotic weight}
  \item{\code{logFm}:}{Numeric scalar. Fishing mortality}
  \item{\code{logA}:}{Numeric scalar. Growth parameter}
  \item{\code{logn}:}{Numeric scalar. Exponent of consumption}
  \item{\code{logeta_F}:}{Numeric scalar. 50\% retention size, relative to asymptotic weight}
  \item{\code{logeta_m}:}{Numeric scalar. 50\% maturation size, relative to asymptotic weight}
  \item{\code{logeta_S}:}{Numeric scalar. 50\% retention size (survey), relative to asymptotic weight}
  \item{\code{loga}:}{Numeric scalar. Physiological mortality}
  \item{\code{logepsilon_a}:}{Numeric scalar. Allocation to activity}
  \item{\code{logepsilon_r}:}{Numeric scalar. Recruitment efficiency}
  \item{\code{logWfs}:}{Numeric scalar. 50\% retention size}
  \item{\code{logu}:}{Numeric scalar. Selectivity parameter, width o}
  \item{\code{M}:}{Numeric scalar. Number of internal weight classes}
  \item{\code{scaleWinf}:}{Numeric scalar. Scale of asymptotic weight}
  \item{\code{scaleFm}:}{Numeric scalar. Scale of fishing mortality}
  \item{\code{scaleA}:}{Numeric scalar. Scale of growth parameter}
  \item{\code{scalen}:}{Numeric scalar. Scale of exponent of consumption}
  \item{\code{scaleeta_F}:}{Numeric scalar. Scale of 50\% retantion size}
  \item{\code{scaleeta_m}:}{Numeric scalar. Scale of 50\% maturation size}
  \item{\code{scaleeta_S}:}{Numeric scalar. Scale of survey gear 50\% retantion size}
  \item{\code{scalea}:}{Numeric scalar. Scale of the physiological mortality}
  \item{\code{scaleepsilon_a}:}{Numeric scalar. Scale of the allocation to activity}
  \item{\code{scaleepsilon_r}:}{Numeric scalar. Scale of recruitment efficiency}
  \item{\code{scaleWfs}:}{Numeric scalar. Scale of size of 50\% retention}
  \item{\code{scaleu}:}{Numeric scalar.}
}
}
\author{
alko
}

