% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment.R
\name{makeAssessment}
\alias{makeAssessment}
\title{Makes assessment of a fish stock given weight frequency data for several years}
\usage{
makeAssessment(inputData, yield = NULL, a.mean = 0.22, a.sd = 0.7,
  nsample = 100, same.as = TRUE, u = 10, sigma = NULL, binsize = NULL,
  winf.ubound = 2, equalWinf = TRUE, probs = seq(0, 1, 0.01),
  seed = as.integer(rnorm(1, 1000, 100)), dirout = "results",
  fnout = format(Sys.time(), "results_\%Y\%m\%d_\%H\%M.RData"), ...)
}
\arguments{
\item{inputData}{list of data.frames, each data.frame has columns \code{Weight} and  \code{Freq} and attribute \code{binsize}}

\item{yield}{numeric, the total yearly catch or landings in kg. Use NULL if not known.}

\item{a.mean}{numeric, physiological mortality.}

\item{a.sd}{numeric, the standard deviation (log domain) of the log-normal distribution of physiological mortality.}

\item{nsample}{integer, number of repetitions for uncertainty etimation. If zero no uncertainty is estimated.}

\item{same.as}{logical, if TRUE use the same random values of physiological mortality for each year.}

\item{u}{numeric, the selectivity steepness parameter.}

\item{sigma}{numeric, if NULL the parameter is estimated, otherwise a constant is used, see Details.}

\item{binsize}{numeric, span of weight classes in grams.}

\item{winf.ubound}{numeric, the upper bound of asymptotic weight. It is a multiplier of the maximum observed weight.}

\item{equalWinf}{logical, if TRUE estimate one asymptotic weight for all years, if FALSE estimate one for each year.}

\item{probs}{numeric vector of probabilites with values in [0,1] for the uncertainty sample quantiles.}

\item{seed}{numeric, the random number generator seed.}

\item{dirout}{Output directory}

\item{fnout}{Output file name}

\item{...}{Arguments passed to \code{estimate_TMB} that does the estimation}
}
\value{
object of class s6modelResults which is a data.frame with parameter estimates with attributes \describe{
\item{CI}{confidence levels as sample quantiles}
\item{obj}{the TMB object of the default value estimation}
\item{opt}{the optimization output from \code{nlminb}}
\item{opts}{the input options used in the run}
\item{Results}{all results using default parameters}
\item{seed}{the random number gernerator used}
\item{timeToCompletion}{difftime object with the time needed for the estimation}
\item{version}{the \code{s6model}version that produced the results}
}
}
\description{
Makes assessment of a fish stock given weight frequency data for several years
}
\details{
\itemize{
\item{sigma paramter} for more information see the cited paper.
}
}
