% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameters-class.R
\docType{class}
\name{Parameters}
\alias{Parameters}
\alias{Parameters-class}
\alias{as.list.Parameters}
\alias{lines.Parameters}
\alias{meanParameters}
\alias{parameters}
\alias{plot.Parameters}
\alias{simulate.Parameters}
\title{The Parameters class and constructor}
\usage{
parameters(pars = list(), base = new("Parameters"))

meanParameters(l)

\method{plot}{Parameters}(x, xlim = c(0.001, 1), ...)

\method{lines}{Parameters}(x, ...)

\method{as.list}{Parameters}(x, ...)

\method{simulate}{Parameters}(object, nsim = 1000, seed = NULL,
  binsize = 100, keepZeros = TRUE, ndataset = 1, ...)
}
\arguments{
\item{pars}{named list, parameter values. If the name starts with log the values are expected to be log transformed}

\item{base}{\code{Parameters} object, the parameters of this object are used instead of the default values}

\item{x}{list of Parameters objects}

\item{xlim}{the x limits (x1, x2) of the plot.}

\item{...}{Arguments passed to other methods.}

\item{object}{a \code{Parameters} object.}

\item{nsim}{number of individuals in the simulated sample.}

\item{seed}{the seed that is passed to \code{\link{set.seed}}.}

\item{binsize}{numeric, the width of the weight classes in grams.}

\item{keepZeros}{logical, if TRUE keep bins with zero frequency.}

\item{ndataset}{integer, the number of datasets to simulate.}

\item{x}{a Parameters object}
}
\value{
A \code{Parameters} object.

\code{meanParameters} returns a \code{Parameters} object with mean 
values of the input Parameters. It returns NULL if \code{x} is NULL and 
\code{x} if \code{x} is an object of class \code{Parameters}
}
\description{
Parameters is an S4 class that contains all model parameters. The 
\code{parameters} function is a constructor of the class. Convienient 
functions are available to \code{plot}, draw \code{lines}, \code{simulate} 
data, return a named list of all parameters (\code{as.list}) and get the 
mean parameters of a list of \code{Parameter} objects.
}
\note{
Additional arguments are passed to \code{\link{plot.default}} (from plot) and to \code{\link{lines}} (from lines). From all other functions the extra arguments are ignored.

In \code{simulate}, all zero bins after the last non-zero bin are droped; even if \code{keepZeros} is TRUE.
}
\section{Slots}{

\describe{
  \item{\code{logWinf}:}{Numeric scalar, asymptotic weight}
  \item{\code{logFm}:}{Numeric scalar, fishing mortality}
  \item{\code{logA}:}{Numeric scalar, growth parameter}
  \item{\code{logn}:}{Numeric scalar, exponent of consumption}
  \item{\code{logeta_F}:}{Numeric scalar, 50\% retention size, relative to asymptotic weight}
  \item{\code{logeta_m}:}{Numeric scalar, 50\% maturation size, relative to asymptotic weight}
  \item{\code{logeta_S}:}{Numeric scalar, 50\% retention size (survey), relative to asymptotic weight}
  \item{\code{loga}:}{Numeric scalar, physiological mortality}
  \item{\code{logepsilon_a}:}{Numeric scalar, allocation to activity}
  \item{\code{logepsilon_r}:}{Numeric scalar, recruitment efficiency}
  \item{\code{logWfs}:}{Numeric scalar, 50\% retention size}
  \item{\code{logu}:}{Numeric scalar, selectivity parameter, width o}
  \item{\code{M}:}{Numeric scalar, number of internal weight classes}
}
}
\examples{

## Without any arguments gives a Parameters object with default values
parameters()

## Log transformed parameters are expected if the parameter name starts with 'log'
par1 <- parameters(c(logWinf = log(1000), logFm = log(0.4), logWfs = log(100)))

## The same is achieved with not transformed parameters
par2 <- parameters(c(Winf = 1000, Fm = 0.4, Wfs = 100))

## Take a Parameters object and change one parameter
par <- parameters(list(Winf = 1000, a = 0.4, Fm = 0.2, Wfs = 100))
changeMatsize <- parameters(list(eta_m = 0.3), base= par)

difference(par, changeMatsize)
##       base comp difference percent.difference
## eta_m 0.25  0.3      -0.05                 20
}
\author{
alko

alko
}
\keyword{constructor}

