% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{plotFit-methods}
\alias{plotFit}
\alias{plotFit-methods}
\title{plotFit}
\usage{
plotFit(object, data, add, ...)
}
\arguments{
\item{object}{A \code{Parameters} object}

\item{data}{Numeric vector or data.frame with columns Weight and Freq.}

\item{add}{Boolean. If TRUE, the plot is added to an existing graphics device.}

\item{...}{Extra named arguments are passed to the plotting function}
}
\value{
invisible NULL
}
\description{
Visualizing fit of s6model
}
\author{
alko
}

