% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotMortality}
\alias{plotMortality}
\title{Makes a plot of natural and fishing mortalities}
\usage{
plotMortality(object, ...)
}
\arguments{
\item{object}{A \code{Parameters} object}

\item{...}{Additional arguments for plot}
}
\value{
Invisible \code{NULL}
}
\description{
Makes a plot of natural and fishing mortalities
}
\author{
alko
}

