% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{plotResiduals}
\alias{plotResiduals}
\title{Make a report about the (quasi-)residuals}
\usage{
plotResiduals(rr)
}
\arguments{
\item{rr}{s6output object, as returned by makeAssessment}
}
\value{
list with residuals and the results of the t-test, acf and Shapiro-Wilks test
}
\description{
Make a report about the (quasi-)residuals
}
