% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment.R
\name{plotSelectivity}
\alias{plotSelectivity}
\title{Plot the selectivity curve(s) from a fitted object}
\usage{
plotSelectivity(res)
}
\arguments{
\item{res}{A fitted s6modelResults object}
}
\value{
Invisible NULL
}
\description{
Plot the selectivity curve(s) from a fitted object
}
\examples{
set.seed(1234)
wfs <- c(1234, 2345, 4567)
f <- 0.2 + cumsum(rnorm(10, 0, 0.01))
s <- c(rep(wfs[1], 3), rep(wfs[3], 4), rep(wfs[2], 3))

dat <- setNames(mapply(function(ff, ss) simulate(parameters(c("Wfs", "Fm"), c(ss, ff), FALSE), binsize = 100), f, s, SIMPLIFY = FALSE),
                2000:2009)
res <- makeAssessment(dat, fnout = NULL, nsample = 1)

plotSelectivity(res)
abline(v = wfs, lty = 2)
mtext(wfs, 3, at = wfs, las = 2)
}
