% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s6input_class.R
\name{s6input}
\alias{s6input}
\alias{s6input-class}
\alias{hist.s6input}
\alias{plot.s6input}
\alias{seq.s6input}
\alias{length.s6input}
\alias{[.s6input}
\alias{format.s6input}
\alias{print.s6input}
\alias{is.s6input}
\title{The s6input class and constructor}
\usage{
s6input(
  wf = NULL,
  surwf = NULL,
  isSimulated = FALSE,
  trueParameters = NULL,
  catch = NULL,
  catchUnits = "tonnes",
  stockname = "-",
  stockcode = "-",
  species = "-",
  verbose = options()$verbose
)

\method{hist}{s6input}(x, ...)

\method{plot}{s6input}(x, y, ...)

\method{seq}{s6input}(x)

\method{length}{s6input}(x)

\method{[}{s6input}(x, i)

\method{format}{s6input}(x, ...)

\method{print}{s6input}(x, ...)

is.s6input(x)
}
\arguments{
\item{wf}{list of data frames with commercial weight frequencies}

\item{surwf}{list of data frames with scientific survey weight frequencies}

\item{isSimulated}{logical, are the data simulated?}

\item{catch}{vector of total catch}

\item{catchUnits}{character, units of catch}

\item{stockname}{character, the stock name, e.g. North Sea cod}

\item{stockcode}{character, the stock code, e.g. cod-347d}

\item{x}{s6input object}

\item{...}{additional paramters to \code{\link{hist}}}

\item{trueParams}{s6params object, the parameters used in the simulation (is used only if isSimulated)}

\item{speciesname}{character, the species latin name, e.g. "Gadus morhua"}
}
\value{
A s6input object
}
\description{
s6input is an S3 class that contains the input data that are used to
make the assessment. The \code{s6input} function is the constructor of the
class. Convenient functions are available to present the data and
\code{plot} them.
}
\examples{
## Empty input
inp <- new_s6input()

## North Sea cod input data
inpall <- s6input(wf = nscoddat, isSurvey = FALSE,
stockcode = "cod.27.47d20",
species = "Gadus morhua",
stockname = "North Sea cod",
years = years,
catch = catch)

## simulate a data set
d <- simulate(s6params(), binsize = 100, nsim = 100)

## Plot the data
hist(d)
}
\author{
Alexandros Kokkalis <alko@aqua.dtu.dk>
}
