% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{s6modelResults}
\alias{plot.s6modelResults}
\alias{print.s6modelResults}
\alias{s6modelResults}
\alias{s6modelResults-class}
\title{s6modelResults class}
\usage{
\method{print}{s6modelResults}(x, ...)

\method{plot}{s6modelResults}(x, ..., what = "FFmsy", use.rownames = FALSE,
  years = NULL, xlab = NULL, ylab = NULL)
}
\description{
s6modelResults class
}

