% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{sample2df}
\alias{sample2df}
\title{Convert a vector sample to data.frame with counts per weight class}
\usage{
sample2df(sam, binsize, keepZeros = TRUE)
}
\arguments{
\item{sam}{Numeric vector. Individual weights}

\item{binsize}{Numeric. Weight class width}

\item{keepZeros}{Logical. If TRUE the resulting data.frame includes weight classes with zero individuals. Otherwise these weight classes are dropped.}
}
\value{
A data.frame with columns Weight and Freq with number per weight class.
}
\description{
Takes a vector containing individual catch weights and returns a data.frame with numbers per weight class
}
\author{
alko
}

