% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{simulateData3}
\alias{simulateData3}
\title{Simulates catch-at-weight data using the s6model}
\usage{
simulateData3(params = s6params(), samplesize = 1000, binsize = 5,
  keepZeros = TRUE, retDF = TRUE, ...)
}
\arguments{
\item{params}{Object of class \code{s6params}. Model parameters}

\item{samplesize}{Integer. Number of simulated individuals}

\item{binsize}{Numeric. Weight class width. If \code{retDF} is FALSE, \code{binsize} is ignored}

\item{keepZeros}{Logical. If TRUE the resulting data.frame includes weight classes with zero individuals. Otherwise these weight classes are dropped. Ignored if \code{retDF} is FALSE}

\item{retDF}{Logical. If TRUE a data.frame is returned with columns Weight (weight classes) and Freq (numbers per weight class)}

\item{...}{Extra named arguments passed to \code{getParams}}
}
\value{
Invisible list with
\itemize{
  \item `sample` containing weights of individuals,
  \item `s6params` the parameters as a \code{s6params} object
  \item `Fmsy` the Fmsy of the given parameters
  \item `df` A data.frame with columns Weight and Freq with number per 
  weight class. This is returned only if \code{retDF} is TRUE
}
}
\description{
Simulate a data set of individual catch weight or
}
\author{
alko
}

