\name{steadyStateEstimation-package}
\alias{steadyStateEstimation-package}
\alias{steadyStateEstimation}
\docType{package}
\title{
Estimates parameters using a steady state size based models
}
\description{
Uses a size-based single species model in steady state to estimate
growth and mortality parameters. 
}
\details{
\tabular{ll}{
Package: \tab steadyStateEstimation\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2012-09-07\cr
License: \tab licence to grill\cr
Depends: \tab methods\cr
Suggests: \tab multicore\cr
}

}
\author{
Alexandros Kokkalis

Maintainer: alko <alko@aqua.dtu.dk>
}

\keyword{ package }

