% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datras.R
\name{subsetDatras}
\alias{subsetDatras}
\title{Subsets a DATRAS object}
\usage{
subsetDatras(dat, species = "Gadus morhua", gear = NULL, years = c(1991,
  2014), haulDur = NULL, icesAreas = NULL, out = NULL)
}
\arguments{
\item{dat}{DATRAS object, or character containing the filename of an RDS file holding a DATRASraw object}

\item{species}{character vector, species to keep (latin name). If \code{NULL} all species are kept}

\item{gear}{character vector, gears to keep. If \code{NULL} all gears are kept}

\item{years}{numeric vector, if length is 2 it is the first and last year to keep, if length is 1 only the data of this year are kept. If \code{NULL} all years are kept}

\item{haulDur}{numeric vector of length 2 with the minimum and maxumum haul duration. If \code{NULL} all haul durations are kept}

\item{icesAreas}{character vector, the ICES area codes to keep. If \code{NULL} all areas are kept}

\item{out}{string, filename of the output RDS file}
}
\value{
if \code{out} is NULL, a DATRASraw object (invisible). If \code{out} is character, the filename of the saved RDS file. If the file specified in \code{out} exist, the function does not rewrite the file and returns the filename.
}
\description{
Selects species, gear, years, haul duration from DATRAS object and returns it
or save it in an RDS file
}
\author{
alko
}
