% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tmclapply}
\alias{tmclapply}
\title{Multicore lapply function with progress bar}
\usage{
tmclapply(X, FUN, ..., progressbar = TRUE)
}
\arguments{
\item{X}{a vector (atomic or list) or an expressions vector.  Other
objects (including classed objects) will be coerced by `as.list`.}

\item{FUN}{the function to be applied to (`mclapply`) each element of `X` or (`mcmapply`) in parallel to `...`.}

\item{...}{Optional arguments to FUN}

\item{progressbar}{Logical. If TRUE a text progress bar is shown.}
}
\value{
Result from mclapply
}
\description{
Wrapper around parallel::mclapply function with text progress bar
}
\note{
If mclapply is not available, lapply is used instead.
}
\author{
alko
}

