% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{Winf<-}
\alias{Winf}
\alias{Winf,Parameters-method}
\alias{Winf<-}
\alias{Winf<-,Parameters-method}
\alias{Winf<--methods,}
\alias{Winfsetter}
\title{Takes a Parameters object and changes its asymptotic weight}
\usage{
Winf(object) <- value

Winf(object)

\S4method{Winf}{Parameters}(object)
}
\arguments{
\item{value}{Numeric. The new asymptotic weight}

\item{object}{\code{Parameters} object}
}
\value{
\code{Parameters} object with changed asymptotic weight, and absolute and
relative 50\% retention sizes
}
\description{
The asymptotic weight is changed, along with the relative and absolute sizes of 50\% retention
}
\author{
alko
}

