% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Winf<-}
\alias{Winf}
\alias{Winf<-}
\title{Takes a Parameters object and changes its asymptotic weight}
\usage{
Winf(object) <- value

Winf(self)
}
\arguments{
\item{object}{\code{Parameters} object}

\item{value}{Numeric. The new asymptotic weight}

\item{self}{\code{Parameters} object}
}
\value{
\code{Parameters} object with changed asymptotic weight, and absolute and
relative 50\% retention sizes
}
\description{
The asymptotic weight is changed, along with the relative and absolute sizes of 50\% retention
}
\author{
alko
}

