% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model.R
\name{calcFmsy}
\alias{calcFmsy}
\title{Calculates the Fmsy reference point}
\usage{
calcFmsy(params = NULL)
}
\arguments{
\item{params}{An object of class \code{Parameters}}
}
\value{
numeric F that leads to MSY
}
\description{
Calculates the Fmsy reference point
}
\author{
alko
}

