% Generated by roxygen2 (4.0.2): do not edit by hand
\name{changeBinsize}
\alias{changeBinsize}
\title{Change the bin size of a weight frequency data.frame}
\usage{
changeBinsize(df, binsize = 10, keepZeros = TRUE)
}
\arguments{
\item{df}{data.frame with colums 'Weight' and 'Freq'}

\item{binsize}{numeric, the bin size in grams}

\item{keepZeros}{logical, if TRUE the bins with zero observation are not removed}
}
\value{
data.frame with weight frequencies binned with bin size \code{binsize}
}
\description{
Change the bin size of a weight frequency data.frame
}
\examples{
## Simulate a data set with bin size equal to 100 gr
dat <- simulateData3(binsize=100)

## Change the bin size to 200 gr
dat <- changeBinsize(dat$df, binsize = 200)
}

