% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{difference}
\alias{difference}
\alias{difference,Parameters,Parameters-method}
\title{Difference between two \code{Parameters} objects}
\usage{
difference(base, comp)

\S4method{difference}{Parameters,Parameters}(base, comp)
}
\arguments{
\item{base}{\code{Parameters} object. First object
##}

\item{comp}{\code{Parameters} object. Second object}
}
\value{
TRUE if they are the same. If there are differences, a data.frame is returned
with the untransformed parameter values of the two objects, the relative difference (base - comp)
and the percent difference
}
\description{
##
}
\author{
alko
}

