% Generated by roxygen2 (4.0.1): do not edit by hand
\name{difference}
\alias{difference}
\title{Difference between two \code{Parameters} objects}
\usage{
difference(base, comp)
}
\arguments{
\item{base}{\code{Parameters} object. First object}

\item{comp}{\code{Parameters} object. Second object}
}
\value{
TRUE if they are the same. If there are differences, a data.frame is returned
with the untransformed parameter values of the two objects, the relative difference (base - comp)
and the percent difference
}
\description{
Difference between two \code{Parameters} objects
}
\author{
alko
}

