% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{estimate}
\alias{estimate}
\title{Estimate fishing mortality and stock status}
\usage{
estimate(inp, params = s6params(...), winf.ubound = 2, est.Winf = TRUE,
  est.Wfs = TRUE, est.etaS = inp@isSurvey, est.a = FALSE,
  est.sigma = TRUE, est.u = FALSE, verbose = FALSE, random = c(),
  make.guess = TRUE, usePois = TRUE, sdloga = 0.7, ...)
}
\arguments{
\item{inp}{\code{\link{s6input}} object containing the input data}

\item{params}{\code{\link{s6params}} object containing the model parameters}

\item{winf.ubound}{numeric, upper bound for Winf}

\item{est.Winf}{logical, the est.* arguments are used to estimate (TRUE) or fix (FALSE) model parameters.}

\item{est.Wfs}{logical}

\item{est.etaS}{logical}

\item{est.a}{logical}

\item{est.sigma}{logical}

\item{est.u}{logical}

\item{verbose}{logical, if \code{TRUE} print TMB output}

\item{random}{character vector, select which parameters should be random effects}

\item{make.guess}{logical, if \code{TRUE} a guess is used to initialise parameter values, otherwise the values from params is used.}

\item{usePois}{logical, if \code{TRUE} (default) the Poisson estimation is used, otherwise a Gaussian estimation is used.}

\item{sdloga}{numeric, the standard deviation of the prior of physiological mortality (a).}

\item{...}{named arguments passed to s6params}
}
\value{
A data.frame with all parameter estimates and derived quantities and their uncertainty.
}
\description{
Function that takes weight frequency distributions and estimates 
the stock status (F/Fmsy), the asymptotic weight and the 50\% retainmen size.
}
\details{
The upper bound of Winf is set using \code{winf.ubound}. The 
bound then is equal to the maximum observed weight in the data times 
\code{winf.ubound}.
}

