% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{estimateParam}
\alias{estimateMultidata}
\alias{estimateParam}
\title{Estimates parameters using the given data and maybe some parameters}
\usage{
estimateParam(names = c("Fm", "Winf", "Wfs"),
  data = simulateData3(parameters(), samplesize = 1000), start = rep(0.5,
  length(names)), lower = rep(-Inf, length(names)), upper = rep(Inf,
  length(names)), fixed.names = c(), fixed.vals = numeric(0),
  fixed.transformed = TRUE, plotFit = FALSE, isSurvey = FALSE,
  verbose = getOption("verbose"), useTMB = TRUE, ...)

estimateMultidata(names = c("Fm", "Winf", "Wfs"),
  surdata = simulateData3(parameters(), samplesize = 1000, isSurvey =
  TRUE)$sample, comdata = simulateData3(parameters(), samplesize =
  1000)$sample, start = rep(0.5, length(names)), lower = rep(0.1,
  length(names)), fixed.names = c(), fixed.vals = numeric(0),
  plotFit = FALSE, ...)
}
\arguments{
\item{names}{String vector. The parameters to be estimated.}

\item{data}{Numeric vector, or \code{data.frame} with columns Weight and Freq, or \code{list} with a numeric vector named `sample` or a \code{data.frame} named `df`.
Weight of individual fish (vector) or frequencies per weight class \code{data.frame}.}

\item{start}{Numeric vector. Initial values of the parameters.}

\item{lower}{The lower bound for the parameter estimation.}

\item{upper}{The upper bound for the parameter estimation.}

\item{fixed.names}{String vector. Names of constants.}

\item{fixed.vals}{Numeric vector. Transformed values of constants.}

\item{fixed.transformed}{Logical. If FALSE the constants are not transformed.}

\item{plotFit}{Boolean. If TRUE a plot is produced with the fited pdf and the
kernel density estimate of the data.}

\item{isSurvey}{Boolean. If TRUE the data are assumed to be from a survey.}

\item{verbose}{Boolean. If TRUE the estimated confidence intervals are printed.}

\item{...}{Additional named arguments passed to plotFit}

\item{surdata}{Same as data. Survey data.}

\item{comdata}{Same as data. Commercial data.}
}
\value{
A Parameters object, containing the estimated parameters.
}
\description{
Parameter estimation minimizing the negative log likelihood, using the
nlminb function. \code{estimateParam} uses data from commercial catches or surveys.
\code{estimateMultidata} uses both sources
}
\note{
If data is a list containing both sample and df, the \code{data.frame} df will be used.
}
\examples{

## Simulate some data
sam <- simulateData3(params=parameters("a", 0.5, transformed=FALSE))

## Estimate the a parameter and see the fitted plot
estimateParam(names="a", data=sam$sample, plotFit=TRUE)
}
\author{
alko
}
\keyword{optimize}

