% Generated by roxygen2 (4.0.1): do not edit by hand
\name{estimateParam}
\alias{estimateMultidata}
\alias{estimateParam}
\title{Estimates parameters using the given data and maybe some parameters}
\usage{
estimateParam(names = c("Fm", "Winf", "Wfs"),
  data = simulateData3(parameters(), samplesize = 1000)$sample,
  start = rep(0.5, length(names)), lower = rep(-Inf, length(names)),
  upper = rep(Inf, length(names)), fixed.names = c(),
  fixed.vals = numeric(0), plotFit = FALSE, isSurvey = FALSE,
  verbose = getOption("verbose"), ...)

estimateMultidata(names = c("Fm", "Winf", "Wfs"),
  surdata = simulateData3(parameters(), samplesize = 1000, isSurvey =
  TRUE)$sample, comdata = simulateData3(parameters(), samplesize =
  1000)$sample, start = rep(0.5, length(names)), lower = rep(0.1,
  length(names)), fixed.names = c(), fixed.vals = numeric(0),
  plotFit = FALSE, ...)
}
\arguments{
\item{names}{String vector. The parameters to be estimated.}

\item{data}{Numeric vector or data.frame with columns Weight and Freq.
Weight of individual fish (vector) or frequencies per weight class (data.frame).}

\item{start}{Numeric vector. Initial values of the parameters.}

\item{lower}{The lower bound for the parameter estimation.}

\item{upper}{The upper bound for the parameter estimation.}

\item{fixed.names}{String vector. Names of constants.}

\item{fixed.vals}{Numeric vector. Transformed values of constants.}

\item{plotFit}{Boolean. If TRUE a plot is produced with the fited pdf and the
kernel density estimate of the data.}

\item{isSurvey}{Boolean. If TRUE the data are assumed to be from a survey.}

\item{verbose}{Boolean. If TRUE the estimated confidence intervals are printed.}

\item{...}{Additional named arguments passed to plotFit}

\item{surdata}{Same as data. Survey data.}

\item{comdata}{Same as data. Commercial data.}
}
\value{
A Parameters object, containing the estimated parameters.

Parameters object with the fitted
}
\description{
Parameter estimation minimizing the negative log likelihood, using the
nlminb function

Similar to estimateParam, but accepting both commercial and survey data
}
\examples{
## Simulate some data
sam <- simulateData3(params=parameters("a", 0.5, transformed=FALSE))

## Estimate the a parameter and see the fitted plot
estimateParam(names="a", data=sam$sample, plotFit=TRUE)
}
\author{
alko

alko
}
\keyword{optimize}

