% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{findLatest}
\alias{findLatest}
\title{Find newest file in folder matching pattern}
\usage{
findLatest(path = ".", pattern = "")
}
\arguments{
\item{path}{character, relative or absolute path}

\item{pattern}{the pattern to search for, used by \code{link{dir}}}
}
\description{
Given a folder and a pattern, rerurns the file that was modified latest.
}
\seealso{
\code{\link{dir}}
}

