% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datras.R
\name{getDfYears}
\alias{getDfYears}
\title{Get yearly weight frequencies}
\usage{
getDfYears(dat, years = as.numeric(levels(dat[[2]]$Year)), binsize = 100,
  ...)
}
\arguments{
\item{dat}{DATRASraw object}

\item{years}{numeric vector, containing the years to return}

\item{binsize}{numeric, the width of the weight classes in grams}

\item{...}{additional arguments to \code{\link{datrasraw2weightfreq}}}
}
\value{
list of \code{data.frame}s, one for each year
}
\description{
For selected years get yearly weight frequencies from DATRAS object
}
\author{
alko
}

