% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getParams}
\alias{getParams}
\title{Returns a list with all parameters and functions}
\usage{
getParams(p = new("Parameters"), FF = NULL, calcBRPs = FALSE,
  isSurvey = FALSE, optim.fmsy = FALSE, optim.fmsyr = FALSE,
  optim.Rrel = FALSE)
}
\arguments{
\item{p}{A Parameters object}

\item{FF}{Numeric. Fishing mortality. This argument is ignored if all optim.* are FALSE}

\item{calcBRPs}{Boolean. If true, calculates biological reference points.}

\item{isSurvey}{boolean, if TRUE a survey selectivity is used for the pdf}

\item{optim.fmsy}{Logical.}

\item{optim.fmsyr}{Logical.}

\item{optim.Rrel}{Logical}
}
\value{
\itemize{
\item If optim.* are all FALSE, an invisible list, containing all model parameters, the cdf, the pdf
functions. spawning stock biomass (SSB), yield. If calcBRPs is TRUE also Fmsy.
\item If optim.fmsy is TRUE, only the yield per Rmax is returned
\item If optim.fmsyr is TRUE, only the yield per recruit is returned
\item If optim.Rrel is TRUE, only the R/Rmax - 0.5 is returned
}
}
\description{
Given a Parameters object, it returns as a list all relevant parameters and
corresponding functions like pdf and cdf.
}
\examples{
getParams()
}
\author{
alko
}
\seealso{
\code{\link{parameters}}
}
\keyword{misc}

