% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimation.R
\name{minimizeme}
\alias{minimizeme}
\alias{minimizemeMultidata}
\title{The negative log likelihood for given parameters and data}
\usage{
minimizeme(theta, data, names, fixed.names = c(), fixed.vals = c(),
  isSurvey = FALSE)

minimizemeMultidata(theta, surdata, comdata, names, fixed.names = c(),
  fixed.vals = c())
}
\arguments{
\item{theta}{Numeric vector of *transformed* parameter values}

\item{data}{Numeric vector, or \code{data.frame} with columns named Weight and Freq. The observed values, fished individuals in grams}

\item{names}{String vector. Contains the names of the parameter vector theta.}

\item{fixed.names}{String vector. Names of constants.}

\item{fixed.vals}{Numeric vector. Transformed values of constants.}

\item{isSurvey}{Logical. If TRUE the observations are assumed to be from a survey.}

\item{surdata}{Same as data. Survey data.}

\item{comdata}{Same as data. Commercial data.}
}
\value{
Numeric scalar. The negative log likelihood for the given parameters and observations.
}
\description{
\code{minimizeme} is mainly used by an optimizer (e.g. \code{optim})
to estimate parameters.
}
\note{
If data is a list containing both sample and df, the \code{data.frame} df will be used.
}
\examples{
\dontrun{
## Simulate some data with default parameter values and fishing mortality Fm = 0.3
sim <- simulateData3(params = parameters("Fm", 0.3, transformed=FALSE))

## Plotting the negative log likelihood for different values of Fm
Fm <- seq(0.1, 1, 0.05)
Fm.transformed <- log(Fm / 0.25)
nll <- sapply(Fm.transformed, function(x) minimizeme(theta = x, data = sim$sample, names = "Fm"))
plot(Fm, nll, type="l")

## Using optimise to estimate one parameter
est <- optimise(f = minimizeme, interval=c(0.01, 1), data=sim$sample, names="Fm")
est.Fm <- exp(est$minimum) * 0.25
abline(v=est.Fm)
mtext(paste("Estimated Fm = ", round(est.Fm, 2)), at=est.Fm)
}
}
\author{
alko
}
\keyword{optimize}

