% Generated by roxygen2 (4.0.1): do not edit by hand
\name{parameters}
\alias{parameters}
\title{Constructor for the Parameters Class}
\usage{
parameters(names = c(), vals = c(), transformed = TRUE,
  base = new("Parameters"))
}
\arguments{
\item{names}{String vector. Contains the names of the parameters that will
have non default values.}

\item{vals}{Numeric vector. The corresponding values, transformed if \code{transformed} is TRUE.}

\item{transformed}{Boolean. If TRUE vals should contain the transformed parameter values.}

\item{base}{\code{Parameters} object. The parameter values will be used instead of the default values.}
}
\value{
Returns an object of the Parameters class
}
\description{
Easier constructor for the Parameters class.
}
\details{
The values in 'vals' should be given as logarithms of the parameter values.
}
\examples{
## Without any arguments gives a Parameters object with default values
parameters()

## Changing some parameters gives the corresponding object
par1 <- parameters(c("Winf", "Fm", "Wfs"), c(log(1000 / 10000), log(0.4 / 0.25), log(100 / 1000)))
par2 <- parameters(c("Winf", "Fm", "Wfs"), c(1000 , 0.4, 100), transformed=FALSE)

## Check if the two objects are equal
all.equal(par1, par2)

## Take a Parameters object and change one parameter
par <- parameters(c("Winf", "a", "Fm", "Wfs"), c(1000, 0.4, 0.2, 100), transformed = FALSE)
changeMatsize <- parameters("eta_m", 0.3, transformed =FALSE, base=par)

difference(par, changeMatsize)
##       base comp difference percent.difference
## eta_m 0.25  0.3      -0.05                 20
}
\author{
alko
}
\keyword{constructor}

