% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s6params-class.R
\docType{methods}
\name{plotFit-methods}
\alias{plotFit}
\alias{plotFit,s6params,data.frame,logical-method}
\alias{plotFit,s6params,data.frame,missing-method}
\alias{plotFit,s6params,numeric,logical-method}
\alias{plotFit,s6params,numeric,missing-method}
\alias{plotFit-methods}
\title{plotFit}
\usage{
plotFit(object, data, add, ...)

\S4method{plotFit}{s6params,numeric,missing}(object, data, add, ...)

\S4method{plotFit}{s6params,data.frame,missing}(object, data, add, ...)

\S4method{plotFit}{s6params,numeric,logical}(object, data, add, ...)

\S4method{plotFit}{s6params,data.frame,logical}(object, data, add, ...)
}
\arguments{
\item{object}{A \code{s6params} object}

\item{data}{Numeric vector or data.frame with columns Weight and Freq.}

\item{add}{Boolean. If TRUE, the plot is added to an existing graphics device.}

\item{...}{Extra named arguments are passed to the plotting function}
}
\value{
invisible \code{NULL}
}
\description{
Visualizing fit of s6model
}

