% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Parameters-class.R
\docType{methods}
\name{plotFit-methods}
\alias{plotFit}
\alias{plotFit,Parameters,data.frame,logical-method}
\alias{plotFit,Parameters,data.frame,missing-method}
\alias{plotFit,Parameters,numeric,logical-method}
\alias{plotFit,Parameters,numeric,missing-method}
\alias{plotFit-methods}
\title{plotFit}
\usage{
plotFit(object, data, add, ...)

\S4method{plotFit}{Parameters,numeric,missing}(object, data, add, ...)

\S4method{plotFit}{Parameters,data.frame,missing}(object, data, add, ...)

\S4method{plotFit}{Parameters,numeric,logical}(object, data, add, ...)

\S4method{plotFit}{Parameters,data.frame,logical}(object, data, add, ...)
}
\arguments{
\item{object}{A \code{Parameters} object}

\item{data}{Numeric vector or data.frame with columns Weight and Freq.}

\item{add}{Boolean. If TRUE, the plot is added to an existing graphics device.}

\item{...}{Extra named arguments are passed to the plotting function}
}
\value{
invisible \code{NULL}
}
\description{
Visualizing fit of s6model


}
\author{
alko
}

