% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{plotGrowth}
\alias{plotGrowth}
\title{Plots growth function}
\usage{
plotGrowth(object, ...)
}
\arguments{
\item{object}{A \code{Parameters} object}

\item{...}{Additional arguments for plot}
}
\value{
Invisible \code{NULL}
}
\description{
Plots growth function
}
\author{
alko
}

