% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Parameters-class.R
\docType{methods}
\name{plotGrowth}
\alias{plotGrowth}
\alias{plotMortality}
\title{Plots growth function}
\usage{
plotGrowth(object, ...)

plotMortality(object, ...)
}
\arguments{
\item{object}{A \code{Parameters} object}

\item{...}{Additional arguments for plot}
}
\value{
Invisible \code{NULL}

Invisible \code{NULL}
}
\description{
Plots growth function

Makes a plot of natural and fishing mortalities
}
\author{
alko

alko
}

