% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s6input-class.R
\docType{class}
\name{s6input}
\alias{[,s6input,ANY,ANY,ANY-method}
\alias{hist.s6input}
\alias{length.s6input}
\alias{s6input}
\alias{s6input-class}
\alias{seq.s6input}
\title{The s6input class and constructor}
\usage{
s6input(wf = list(), surWF = list(), isSimulated = FALSE,
  trueParams = s6params(), isSurvey = FALSE, catch = numeric(0),
  years = numeric(0), stockname = "generic stock", stockcode = "-",
  speciesname = "-")

\method{hist}{s6input}(x, ...)

\S4method{[}{s6input,ANY,ANY,ANY}(x, i)

\method{seq}{s6input}(x)

\method{length}{s6input}(x)
}
\arguments{
\item{wf}{list of data frames with commercial weight frequencies}

\item{surWF}{list of data frames with scientific survey weight frequencies}

\item{isSimulated}{logical, are the data simulated?}

\item{trueParams}{s6params object, the parameters used in the simulation (is used only if isSimulated)}

\item{isSurvey}{logical, if TRUE the surWF is used and wf is ignored and vice versa}

\item{catch}{vectors of total catch in tonnes}

\item{years}{vector of years}

\item{stockname}{character, the stock name, e.g. North Sea cod}

\item{stockcode}{character, the stock code, e.g. cod-347d}

\item{speciesname}{character, the species latin name, e.g. "Gadus morhua"}

\item{x}{a s6input object}

\item{...}{additional paramters to \code{\link{hist}}}
}
\value{
A s6input object

Nothing, it is just used for to plot the
}
\description{
s6input is an S4 class that contains the input data that are used to 
make the assessment. The \code{s6input} function is the constructor of the
class. Convienient functions are available to present the data and 
\code{plot} them.

Plot the weight distribution
}
\section{Slots}{

\describe{
\item{\code{wf}}{list.}

\item{\code{surWF}}{list.}

\item{\code{isSimulated}}{logical.}

\item{\code{trueParams}}{s6params.}

\item{\code{isSurvey}}{logical.}

\item{\code{catch}}{numeric.}

\item{\code{years}}{numeric.}

\item{\code{stockname}}{chacharacter.}

\item{\code{stockcode}}{character.}

\item{\code{speciesname}}{character.}
}}
\examples{
## simulate a data set
d <- simulate(s6params(), binsize = 100, nsim = 100)

## Plot the data
hist(d)
}
\author{
alko
}

