% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s6input_class.R
\name{s6input}
\alias{s6input}
\alias{s6input-class}
\alias{hist.s6input}
\alias{seq.s6input}
\alias{length.s6input}
\alias{[.s6input}
\alias{format.s6input}
\alias{print.s6input}
\title{The s6input class constructor}
\usage{
s6input(
  wf = NULL,
  surwf = NULL,
  isSimulated = FALSE,
  trueParameters = NULL,
  isSurvey = FALSE,
  catch = NULL,
  catchUnits = "tonnes",
  years = NULL,
  stockname = "-",
  stockcode = "-",
  species = "-"
)

\method{hist}{s6input}(x, ...)

\method{seq}{s6input}(x)

\method{length}{s6input}(x)

\method{[}{s6input}(x, i)

\method{format}{s6input}(x, ...)

\method{print}{s6input}(x)
}
\arguments{
\item{wf}{list of data frames with commercial weight frequencies}

\item{surwf}{list of data frames with scientific survey weight frequencies}

\item{isSimulated}{logical, are the data simulated?}

\item{isSurvey}{logical, if TRUE the surWF is used and wf is ignored and vice versa}

\item{catch}{vector of total catch}

\item{catchUnits}{character, units of catch}

\item{years}{vector of years}

\item{stockname}{character, the stock name, e.g. North Sea cod}

\item{stockcode}{character, the stock code, e.g. cod-347d}

\item{x}{s6input object}

\item{...}{additional paramters to \code{\link{hist}}}

\item{trueParams}{s6params object, the parameters used in the simulation (is used only if isSimulated)}

\item{speciesname}{character, the species latin name, e.g. "Gadus morhua"}
}
\value{
A s6input object

Nothing, it is just used for to plot the
}
\description{
s6input is an S3 class that contains the input data that are used to 
make the assessment. The \code{s6input} function is the constructor of the
class. Convenient functions are available to present the data and 
\code{plot} them.
}
\examples{
 
## Empty input
inp <- new_s6input()

## simulate a data set
d <- simulate(s6params(), binsize = 100, nsim = 100)

## Plot the data
hist(d)
}
\author{
Alexandros Kokkalis <alko@aqua.dtu.dk>
}
