% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s6input-class.R
\docType{class}
\name{s6input}
\alias{hist.s6input}
\alias{s6input}
\alias{s6input-class}
\title{The s6input class and constructor}
\usage{
\method{hist}{s6input}(x, ...)
}
\arguments{
\item{x}{a s6input object}

\item{...}{additional paramters to \code{\link{hist}}}
}
\value{
Nothing, it is just used for to plot the
}
\description{
s6input is an S4 class that contains the input data that are used to 
make the assessment. The \code{s6input} function is the constructor of the
class. Convienient functions are available to present the data and 
\code{plot} them.

Plot the weight distribution
}
\section{Slots}{

\describe{
\item{\code{wf}}{list.}

\item{\code{surWF}}{list.}

\item{\code{isSimulated}}{logical.}

\item{\code{trueParams}}{s6params.}

\item{\code{isSurvey}}{logical.}

\item{\code{catch}}{numeric.}

\item{\code{years}}{numeric.}

\item{\code{stockname}}{chacharacter.}

\item{\code{stockcode}}{character.}

\item{\code{speciesname}}{character.}
}}
\examples{
## simulate a data set
d <- simulate(s6params(), binsize = 100, nsim = 100)

## Plot the data
hist(d)
}
\author{
alko
}

