% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{s6model}
\alias{s6model}
\alias{s6model-package}
\title{Data poor stock assessment using a single species size based model in steady state}
\description{
Data poor stock assessment using a single species size based model in steady state
}
\references{
Kokkalis, A., Nielsen, A., Thygesen U. H. and K. H. Andersen. (In review) Reliability of fisheries reference points estimation for data-poor stocks, Fisheries research

Andersen, K. H., & Beyer, J. E. (2013). Size structure, not metabolic scaling rules, determines fisheries reference points. Fish and Fisheries. doi:10.1111/faf.12042
}

