% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{s6modelResults}
\alias{plot.s6modelResults}
\alias{print.s6modelResults}
\alias{s6modelResults}
\alias{s6modelResults-class}
\title{s6modelResults class}
\usage{
\method{print}{s6modelResults}(x, ...)

\method{plot}{s6modelResults}(x, ..., what = "FFmsy", use.rownames = FALSE,
  years = NULL, xlab = NULL, ylab = NULL, ylim = NULL,
  addDefault = FALSE, col.def = "white", addhline = 1, col.hline = 1,
  lty.hline = 2, addMedian = FALSE, col.median = "white",
  lty.median = 2, cex.ver = 0.7, version = TRUE, xaxs = "i",
  yaxs = "i", ci = c("bootstrap", "estimated"), grey.intensity = 1)
}
\description{
s6modelResults class
}

