% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s6params_class.R
\name{s6params}
\alias{s6params}
\alias{s6params-class}
\alias{print.s6params}
\alias{is.s6params}
\alias{all.equal.s6params}
\title{The s6params class and constructor}
\usage{
s6params(
  Winf = 10000,
  Fm = 0.25,
  A = 4.47,
  n = 0.75,
  eta_m = 0.25,
  eta_S = 0.001,
  a = 0.22,
  epsilon_a = 0.8,
  epsilon_r = 0.1,
  Wfs = 500,
  u = 10,
  ngrid = 1000,
  wl.a = 0.01,
  wl.b = 3,
  base = NULL
)

\method{print}{s6params}(x, ...)

is.s6params(x)

\method{all}{equal.s6params}(target, current)
}
\arguments{
\item{Winf}{numeric, asymptotic weight}

\item{Fm}{numeric, fishing mortality}

\item{A}{numeric, growth parameter}

\item{n}{numeric, exponent of consumption}

\item{eta_m}{numeric, 50\% maturation size, relative to asymptotic weight}

\item{eta_S}{numeric, 50\% retention size (survey), relative to asymptotic weight}

\item{a}{numeric, physioical mortality}

\item{epsilon_a}{numeric, allocation to activity}

\item{epsilon_r}{numeric, recruitment efficiency}

\item{Wfs}{numeric, 50\% retention size}

\item{u}{numeric, selectivity parameter}

\item{ngrid}{numeric, number of internal weight classes for Fmsy estimation}

\item{wl.a, }{numeric, weight lenght relationship multiplier}

\item{wl.b}{numeric, weight lenght relationship exponent}

\item{base}{s6params object}

\item{eta_F}{numeric, 50\% retention size, relative to asymptotic weight}
}
\description{
s6params is an S3 class that contains all model parameters. The 
\code{s6params} function is the constructor of the class. Convienient 
functions are available to \code{plot}, draw \code{lines}, \code{simulate} 
data, return a named list of all parameters (\code{as.list}) and get the 
mean parameters of a list of \code{s6params} objects.
}
\examples{
## Without any arguments gives a \code{s6params} object with default values
s6params()

## An object with some user selected values
par1 <- s6params(Winf = 1000, Fm = 0.4, Wfs = 100)
 
## Take a s6params object and change one parameter
par <- s6params(Winf = (1000), a = (0.4), Fm = (0.2), Wfs = (100))
changeMatsize <- par
changeMatsize$etam <- (0.3)

difference(par, changeMatsize)
##       base comp difference percent.difference
## eta_m 0.25  0.3      -0.05                 20
}
\author{
Alexandros Kokkalis <alko@aqua.dtu.dk>
}
