% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s6params-class.R
\docType{class}
\name{s6params}
\alias{as.list.s6params}
\alias{lines.s6params}
\alias{meanParameters}
\alias{plot.s6params}
\alias{s6params}
\alias{s6params-class}
\alias{simulate.s6params}
\title{The s6params class and constructor}
\usage{
s6params(pars = list(), base = new("s6params"))

meanParameters(l, ...)

\method{plot}{s6params}(x, xlim = c(0.001, 1), ...)

\method{lines}{s6params}(x, ...)

\method{as.list}{s6params}(x, ...)

\method{simulate}{s6params}(object, nsim = 1000, seed = NULL,
  binsize = 100, keepZeros = TRUE, ndataset = 1, ...)
}
\arguments{
\item{pars}{named list, parameter values. If the name starts with log the values are expected to be log transformed}

\item{base}{\code{s6params} object, the parameters of this object are used instead of the default values}

\item{l}{list of \code{s6params} objects}

\item{...}{Arguments passed to other methods.}

\item{xlim}{the x limits (x1, x2) of the plot.}

\item{object}{a \code{s6params} object.}

\item{nsim}{number of individuals in the simulated sample.}

\item{seed}{the seed that is passed to \code{\link{set.seed}}.}

\item{binsize}{numeric, the width of the weight classes in grams.}

\item{keepZeros}{logical, if TRUE keep bins with zero frequency.}

\item{ndataset}{integer, the number of datasets to simulate.}

\item{l}{a s6params object}
}
\value{
A \code{s6params} object.

A \code{s6params} object with mean 
values of the input \code{s6params}. If the input is NULL it returns NULL
and it returns the i

It returns NULL if \code{x} is NULL and 
\code{l} if \code{l} is an object of class \code{s6params}
}
\description{
s6params is an S4 class that contains all model parameters. The 
\code{s6params} function is a constructor of the class. Convienient 
functions are available to \code{plot}, draw \code{lines}, \code{simulate} 
data, return a named list of all parameters (\code{as.list}) and get the 
mean parameters of a list of \code{s6params} objects.

Calculate the mean value of
}
\note{
Additional arguments are passed to \code{\link{plot.default}} 
(from plot) and to \code{\link{lines}} (from lines). From all other 
functions the extra arguments are ignored.

In \code{simulate}, all zero bins after the last non-zero bin are droped; even if \code{keepZeros} is TRUE.
}
\section{Slots}{

\describe{
  \item{\code{logWinf}:}{Numeric scalar, asymptotic weight}
  \item{\code{logFm}:}{Numeric scalar, fishing mortality}
  \item{\code{logA}:}{Numeric scalar, growth parameter}
  \item{\code{logn}:}{Numeric scalar, exponent of consumption}
  \item{\code{logeta_F}:}{Numeric scalar, 50\% retention size, relative to asymptotic weight}
  \item{\code{logeta_m}:}{Numeric scalar, 50\% maturation size, relative to asymptotic weight}
  \item{\code{logeta_S}:}{Numeric scalar, 50\% retention size (survey), relative to asymptotic weight}
  \item{\code{loga}:}{Numeric scalar, physiological mortality}
  \item{\code{logepsilon_a}:}{Numeric scalar, allocation to activity}
  \item{\code{logepsilon_r}:}{Numeric scalar, recruitment efficiency}
  \item{\code{logWfs}:}{Numeric scalar, 50\% retention size}
  \item{\code{logu}:}{Numeric scalar, selectivity parameter, width o}
  \item{\code{M}:}{Numeric scalar, number of internal weight classes}
}
}
\examples{

## Without any arguments gives a \\code{s6params} object with default values
s6params()

## Log transformed s6params are expected if the parameter name starts with 'log'
par1 <- s6params(c(logWinf = log(1000), logFm = log(0.4), logWfs = log(100)))

## The same is achieved with not transformed parameters
par2 <- s6params(c(Winf = 1000, Fm = 0.4, Wfs = 100))

## Take a s6params object and change one parameter
par <- s6params(list(Winf = 1000, a = 0.4, Fm = 0.2, Wfs = 100))
changeMatsize <- s6params(list(eta_m = 0.3), base= par)

difference(par, changeMatsize)
##       base comp difference percent.difference
## eta_m 0.25  0.3      -0.05                 20
}
\author{
alko

alko
}
\keyword{constructor}

