% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s6params-class.R
\docType{methods}
\name{Winf<-}
\alias{Winf}
\alias{Winf,s6params-method}
\alias{Winf<-}
\alias{Winf<-,s6params-method}
\alias{Winf<--methods,}
\alias{Winfsetter}
\title{Takes a \code{s6params} object and changes its asymptotic weight}
\usage{
Winf(object) <- value

\S4method{Winf}{s6params}(object) <- value

Winf(object)

\S4method{Winf}{s6params}(object)
}
\arguments{
\item{object}{\code{s6params} object}

\item{value}{Numeric. The new asymptotic weight}
}
\value{
\code{s6params} object with changed asymptotic weight, and absolute and
relative 50\% retention sizes
}
\description{
The asymptotic weight is changed, along with the relative and absolute sizes of 50\% retention
}
\author{
alko
}

